+++ yeat/cli/__init__.py
@@ -7,4 +7,4 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
-from .cli import main
+from .cli import *
