+++ yeat/workflows/workflows.py
@@ -7,12 +7,69 @@
 # Development Center.
 # -------------------------------------------------------------------------------------------------
 
-from . import short, long
+from pathlib import Path
+from pkg_resources import resource_filename
+from snakemake import snakemake
 
 
-def run_shortread_workflow(args, assembly_configs):
-    if args.paired:
-        short.run_paired(
+def run_paired(
+    fastq1,
+    fastq2,
+    assembly_configs,
+    outdir=".",
+    cores=1,
+    sample="sample",
+    dryrun="dry",
+    downsample=0,
+    coverage=150,
+    seed=None,
+    genomesize=0,
+):
+    snakefile = resource_filename("yeat", "workflows/snakefiles/Paired")
+    r1 = Path(fastq1).resolve()
+    if not r1.is_file():
+        raise FileNotFoundError(f"No such file: '{r1}'")
+    r2 = Path(fastq2).resolve()
+    if not r2.is_file():
+        raise FileNotFoundError(f"No such file: '{r2}'")
+    assemblers = [config.algorithm for config in assembly_configs]
+    extra_args = {config.algorithm: config.extra_args for config in assembly_configs}
+    config = dict(
+        read1=r1,
+        read2=r2,
+        assemblers=assemblers,
+        extra_args=extra_args,
+        sample=sample,
+        downsample=downsample,
+        coverage=coverage,
+        seed=seed,
+        genomesize=genomesize,
+    )
+    success = snakemake(
+        snakefile, config=config, cores=cores, dryrun=dryrun, printshellcmds=True, workdir=outdir
+    )
+    if not success:
+        raise RuntimeError("Snakemake Failed")
+
+
+def run_pacbio(fastq, assembly_configs, outdir=".", cores=1, sample="sample", dryrun="dry"):
+    snakefile = resource_filename("yeat", "workflows/snakefiles/Pacbio")
+    fastq = Path(fastq).resolve()
+    if not fastq.is_file():
+        raise FileNotFoundError(f"No such file: '{fastq}'")
+    assemblers = [config.algorithm for config in assembly_configs]
+    extra_args = {config.algorithm: config.extra_args for config in assembly_configs}
+    config = dict(fastq=fastq, assemblers=assemblers, extra_args=extra_args, sample=sample)
+    success = snakemake(
+        snakefile, config=config, cores=cores, dryrun=dryrun, printshellcmds=True, workdir=outdir
+    )
+    if not success:
+        raise RuntimeError("Snakemake Failed")
+
+
+def run_workflow(args, assembly_configs):
+    if args.paired is not None:
+        run_paired(
             *args.paired,
             assembly_configs=assembly_configs,
             outdir=args.outdir,
@@ -24,13 +81,8 @@ def run_shortread_workflow(args, assembly_configs):
             seed=args.seed,
             genomesize=args.genome_size,
         )
-    # elif args.single:
-    # elif args.interleaved:
-
-
-def run_longread_workflow(args, assembly_configs):
-    if args.pacbio:
-        long.run_pacbio(
+    elif args.pacbio is not None:
+        run_pacbio(
             args.pacbio,
             assembly_configs=assembly_configs,
             outdir=args.outdir,
@@ -38,12 +90,3 @@ def run_longread_workflow(args, assembly_configs):
             sample=args.sample,
             dryrun=args.dry_run,
         )
-    # elif args.single:
-    # elif args.interleaved:
-
-
-def run_workflow(args, assembly_configs):
-    if args.readtype == "short":
-        run_shortread_workflow(args, assembly_configs)
-    elif args.readtype == "long":
-        run_longread_workflow(args, assembly_configs)
